feat: add question status icon in question detail
